add Signal with metadata & scripted auto-download APK updates
The app metadata needs to be maintained in this git repo since there
is not currently a canonical source from upstream.  So this adds
metadata and screenshots and icon.

The script downloads the most recent Signal APK if there is a newer
release available.  This script is designed to be run frequently in a
cron job, even as often as once a minute.

Change-Id: Ie9c7dee93e16027d7e44ae868a5069bd42111e87

https: //gitlab.com/CalyxOS/calyxos/-/issues/490
Change-Id: Ieb0ef32bf6bbdec24e26c08fdf2e097394cde4d4