hack until app ID scheme is worked out in platform_prebuilts_calyx_fdroid
In order to change the platform_prebuilts_calyx_fdroid repo as little as
possible, this is a hack to force the APKs to be named following the scheme
that is currently used there.  The ./check-* scripts can't just write out
to that naming scheme because this whole system relies on `fdroid update`
to strip down the number of APKs per app to one in a way that also respects
the ABIs included in the APK.  For that to work, there has to be multiple
APKs per app in repo/.  The platform_prebuilts_calyx_fdroid naming scheme
only uses the app name.

Change-Id: I3c03d4ba95cdb7cd6e902a7769c74158d28c99de